Sync Fork from Upstream Repo

+++ HostnameMeter.c
@@ -19,9 +19,9 @@ static const int HostnameMeter_attributes[] = {
    HOSTNAME
 };
 
-static void HostnameMeter_updateValues(Meter* this, char* buffer, int size) {
+static void HostnameMeter_updateValues(Meter* this, char* buffer, size_t size) {
    (void) this;
-   gethostname(buffer, size-1);
+   gethostname(buffer, size - 1);
 }
 
 const MeterClass HostnameMeter_class = {
